requirements.txt
protobuf
aiogram==3.13.1
httpx==0.27.2
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.35.0
loguru==0.7.2